Revert "chkoutall&reinstall (local)"
This reverts commit d4b9dd42fc008ac55d917f4e98e24d5bf0a64591.

+++ src/openne/hope.py
@@ -31,7 +31,7 @@ class HOPE(object):
 
         self._beta = 0.1 # 0.0728
 
-        M_g = torch.eye(graph.number_of_nodes()) - self._beta * A
+        M_g = torch.eye(graph.number_of_nodes(),dtype=torch.float64) - self._beta * A
         M_l = self._beta * A
 
         # M_g = np.eye(graph.number_of_nodes())
